refactor : rename TextField to Test
